Make new framework for audio streams to enable transformations in a manageable way, and use it to add changing playback speed.

+++ file.go
@@ -31,7 +31,7 @@ type File struct {
 		Text string
 		Bar  int
 	}
-	Volume     float64
+	Volume     int
 	PitchShift int
 
 	dir string
